#52993 ParametersCallbackMap ignores a null value if a variable was already defined
Fixed ParametersCallbackMap(PCM) to always use an overridden value, even if it's null. Updated unit tests. 
In child script/query elements, PCM treats variable=null as undefined and this cause an error in JavaScript. The following check has to be used to workaround this: i = (typeof i == 'undefined')? null : i;