Add source and target to parent project.
I added source and target settings for the Maven compiler to indicate
the target version level of the software.  There are not many changes
necessary to get scriptella to JDK8 compliance, but indicating the
current version will allow that support to be specifically targeted
as a goal for the project while fixing compile errors that occur in
eclipse in the meantime.